chore: move snapshot to open action
