+++ apps/web/src/components/Shared/OpenAction/Snapshot/Choices.tsx
@@ -20,7 +20,7 @@ import useHandleWrongNetwork from 'src/hooks/useHandleWrongNetwork';
 import { useAppStore } from 'src/store/app';
 import { useSignTypedData } from 'wagmi';
 
-import New from '../Badges/New';
+import New from '../../Badges/New';
 import VoteProposal from './VoteProposal';
 
 interface ChoicesProps {
