+++ apps/web/src/components/Shared/OpenAction/Snapshot/Header.tsx
