+++ apps/web/src/components/Shared/OpenAction/Snapshot/VoteProposal.tsx
